Update and rename Day_17_Extended_version_of_day_14 to Day_18_Guess_the_number
We are going to build a "Guess the Number" guessing game.

You are going to use a while loop and some of the concepts from the past few days.

Start by picking a number between 0 and 1,000,000. This will be your first variable.
Create a while loop to keep asking the user to guess your number.

If they are too low, tell them "too low." If they guess too high tell, them "too high."

If the user guesses correctly, tell them they are a winner (maybe add a fun emoji too!)
Count the number of attempts it took for the user to guess number. Make sure you only show that after they get the answer correct.
Extra challenge: If the user types a negative number, exit program.