Update Day_27_first_video_game_character_creation
Welcome to your first video game creation. You will create a video game that creates a character's health and attack stats...along with an epic name for your character.

Do not delete today's code. You will be building upon it on Day 28.

Write a subroutine that generates a character: first name and character type (human, imp, wizard, elf, etc.).
Write a subroutine that multiplies a bunch of random dice rolls together to generate the character's health stats. Use this formula:


Write a second subroutine that multiplies a bunch of random dice rolls together to generate the character's strength stats. Use this formula:


Present the data in a menu with time.sleep and os.system("clear") to make it appealing.
Wrap it in a loop so the user has the option to create another character.