Create Day_42_Dictionary_Mokebeast.py
Some trainers have no fear. To them, this is just one more challenge".

Create a dictionary to store the details of your, ahem, MokéBeast.
Ask the user to input the following details: name, type (earth, fire, air, water or spirit), special move, starting HP and starting MP. For now we're just taking in one set of values for one beast.
Output the beast's details.
Check the beast's type and change the color of the text accordingly. Fire is red, water is blue, air is white. You decide on the others.
🥳 Extra points for getting all the inputs with just one input command and the split function.